Refactor: 변경된 util 함수 이름으로 import 수정
